+++ xarray/core/dask_array_ops.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import math
+
 from xarray.core import dtypes, nputils
 
 
@@ -29,6 +31,15 @@ def dask_rolling_wrapper(moving_func, a, window, min_count=None, axis=-1):
 def least_squares(lhs, rhs, rcond=None, skipna=False):
     import dask.array as da
 
+    from xarray.core.dask_array_compat import reshape_blockwise
+
+    if rhs.ndim > 2:
+        out_shape = rhs.shape
+        reshape_chunks = rhs.chunks
+        rhs = reshape_blockwise(rhs, (rhs.shape[0], math.prod(rhs.shape[1:])))
+    else:
+        out_shape = None
+
     lhs_da = da.from_array(lhs, chunks=(rhs.chunks[0], lhs.shape[1]))
     if skipna:
         added_dim = rhs.ndim == 1
@@ -52,6 +63,17 @@ def least_squares(lhs, rhs, rcond=None, skipna=False):
         # Residuals here are (1, 1) but should be (K,) as rhs is (N, K)
         # See issue dask/dask#6516
         coeffs, residuals, _, _ = da.linalg.lstsq(lhs_da, rhs)
+
+    if out_shape is not None:
+        coeffs = reshape_blockwise(
+            coeffs,
+            shape=(coeffs.shape[0], *out_shape[1:]),
+            chunks=((coeffs.shape[0],), *reshape_chunks[1:]),
+        )
+        residuals = reshape_blockwise(
+            residuals, shape=out_shape[1:], chunks=reshape_chunks[1:]
+        )
+
     return coeffs, residuals